+++ ui/src/scenes/Home/components/NeedsContainer/components/NeedsList/components/MissingRealizersAmount/MissingRealizersAmount.js
@@ -1,0 +1,30 @@
+import React from 'react';
+import styled from 'styled-components';
+import RealizersMissingIcon from 'components/RealizersMissingIcon';
+
+const RightMarginSpan = styled.span`
+  margin-right: 10px;
+`;
+
+const MissingRealizersAmount = ({ need }) => {
+  let realizersMissing = [];
+  if (need.fulfilledBy) {
+    realizersMissing = need.fulfilledBy.filter((resp) => !resp.realizer);
+  }
+
+  if (realizersMissing.length > 0) {
+    return (
+      <div>
+        <RightMarginSpan>
+          {realizersMissing.length}
+          x
+        </RightMarginSpan>
+        {' '}
+        <RealizersMissingIcon />
+      </div>
+    );
+  }
+  return '';
+};
+
+export default MissingRealizersAmount;
