+++ ui/src/scenes/Home/components/NeedsContainer/components/NeedsList/components/NeedsListItem/index.js
@@ -1,0 +1,3 @@
+import NeedsListItem from './NeedsListItem';
+
+export default NeedsListItem;
